Refactor zipcontent backend to run from separate wsgi server.
